pyproject.toml
[project]
name = "jwt-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "bazuly"}
]
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "make (>=0.1.6.post2,<0.2.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-json-logger (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
